% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseRange.R
\name{ChooseRange}
\alias{ChooseRange}
\title{Choose range for the refined cross PCF ratio estimator}
\usage{
ChooseRange(X, covariate, Beta, r, bwd, kern = "Epanechnikov", cPCF = NULL)
}
\arguments{
\item{X}{A multivariate point process. X must be of class ppp.}

\item{covariate}{Covariates. The covariates must be a matrix. The rows corresponds to the points in the point pattern and the
columns indicates the corresponding covariates that are observed at the location of the point pattern.}

\item{Beta}{Estimated first order parameters. Beta must be a matrix. The number of rows must correspond to the number of covariates.
The number of columns must correspond to the number of point types.}

\item{r}{A vector of distances}

\item{bwd}{The bandwidth used}

\item{kern}{The kernel function. Default is Epanechnikov. Alternatively, the kernel function can be Indicator.}

\item{cPCF}{Non parametric estimates of cross PCF ratios. Default is cPCF = NULL. If cPCF = NULL, then the cross PCF ratios are estimated
using the function CrossPCF}
}
\value{
Return estimated cross PCF ratios, r distances used, bandwidth used, refined cross PCF ratios and R^*
}
\description{
This is a method to choose the range for the refined cross PCF ratio estimator
}
\references{
Hessellund, K. B., Xu, G., Guan, Y. and Waagepetersen, R. (2020)
Second order semi-parametric inference for multivariate log Gaussian Cox processes.
}
\author{
Kristian Bjørn Hessellund, Ganggang Xu, Yongtao Guan and Rasmus Waagepetersen.
}
