% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimedata.R
\docType{data}
\name{markedprocess}
\alias{markedprocess}
\title{Washington D.C. street crimes in January-February 2017}
\format{
A multivariate point pattern data, where each type of point correspond to a specific type of street crime committed in Wastington D.C.
\describe{
\item{window}{observation window}
\item{n}{number of data points}
\item{x}{x coordinate}
\item{y}{y coordinate}
\item{markformat}{vector}
\item{marks}{marks to each data point}
}
}
\source{
http://opendata.dc.gov/datasets/crime-incidents-in-2017
}
\usage{
markedprocess
}
\description{
Washington D.C. street crimes in January-February 2017
}
\examples{
markedprocess

}
\references{
Hessellund, K. B., Xu, G., Guan, Y. and Waagepetersen, R. (2020)
Semi-parametric multinomial logistic regression for multivariate point pattern data.
}
\author{
Kristian Bjørn Hessellund, Ganggang Xu, Yongtao Guan and Rasmus Waagepetersen.
}
\keyword{datasets}
