% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_lik.R
\name{second_order_likelihood}
\alias{second_order_likelihood}
\title{Second order conditional composite likelihood function}
\usage{
second_order_likelihood(
  psi,
  xis,
  sigmas,
  phis,
  R,
  fi,
  fj,
  mi,
  mj,
  mval,
  stz = F
)
}
\arguments{
\item{psi}{Current value of alpha}

\item{xis}{Current value of xi (log transformed)}

\item{sigmas}{Current value of sigma^2 (log transformed)}

\item{phis}{Current value of phi (log transformed)}

\item{R}{distance}

\item{fi}{exp(beta^T z(u))}

\item{fj}{exp(beta^T z(v))}

\item{mi}{mark of u}

\item{mj}{mark of v}

\item{mval}{all marks}

\item{stz}{indicator}
}
\value{
likelihood
}
\description{
Second order conditional composite likelihood function.
}
