% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_sigma2.R
\name{sensitivity.sigma2}
\alias{sensitivity.sigma2}
\title{Sensitivity of sigma^2}
\usage{
sensitivity.sigma2(
  alphas,
  xis,
  sigmas,
  phis,
  R = 0,
  fi,
  fj,
  mi = NULL,
  mj = NULL,
  mval = NULL
)
}
\arguments{
\item{alphas}{Current value of alpha}

\item{xis}{Current value of xi (log transformed)}

\item{sigmas}{Current value of sigma^2 (log transformed)}

\item{phis}{Current value of phi (log transformed)}

\item{R}{distance}

\item{fi}{exp(beta^T z(u))}

\item{fj}{exp(beta^T z(v))}

\item{mi}{mark of u}

\item{mj}{mark of v}

\item{mval}{all marks}
}
\value{
Sensitivity of sigma^2
}
\description{
Sensitivity of sigma^2
}
