% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\docType{data}
\name{Z_process}
\alias{Z_process}
\title{Covariates for Washington D.C. street crimes in January-February 2017}
\format{
A matrix of spatial covariates. The rows corresponds to the spatial covariates
observed at the street crime data points. The columns indicates the spatial covariates.
}
\source{
http://opendata.dc.gov/datasets
}
\usage{
Z_process
}
\description{
Covariates for Washington D.C. street crimes in January-February 2017
}
\examples{
Z_process

}
\references{
Hessellund, K. B., Xu, G., Guan, Y. and Waagepetersen, R. (2020)
Semi-parametric multinomial logistic regression for multivariate point pattern data.
}
\author{
Kristian Bjørn Hessellund, Ganggang Xu, Yongtao Guan and Rasmus Waagepetersen.
}
\keyword{covariates}
\keyword{datasets,}
